feat(consolidation): unify HTTP client and URL utilities
T22: HTTP Client Duplication Elimination (-146 LOC)
- Merge network/client.py and core/client.py into unified implementation
- Add CachedHttpClient, context manager support, batch operations
- Enhanced connection pooling, retry logic, and error handling
- Remove entire network/ module to eliminate duplication

T24: URL Utilities Consolidation (-104 LOC)
- Create comprehensive url_utils.py with 9 utility functions
- Consolidate filename generation, validation, domain extraction
- Eliminate duplicate URL processing across multiple modules
- Add type hints, documentation, and comprehensive validation

Code quality improvements:
- Single source of truth for HTTP operations
- Centralized URL processing with reusable functions
- Enhanced error handling with structured context
- Comprehensive documentation and examples

Related tasks: T22, T24
