--- markdown_lab/network/__init__.py
@@ -1,1 +1,0 @@
-"""Network utilities for markdown_lab."""
