+++ markdown_lab/utils/__init__.py
@@ -1,1 +1,25 @@
 """Utility modules for Markdown Lab."""
+
+from .url_utils import (
+    extract_base_url,
+    get_domain_from_url,
+    get_filename_from_url,
+    get_url_path_parts,
+    is_absolute_url,
+    normalize_url,
+    parse_url_safe,
+    sanitize_filename_part,
+    validate_url,
+)
+
+__all__ = [
+    "extract_base_url",
+    "get_domain_from_url", 
+    "get_filename_from_url",
+    "get_url_path_parts",
+    "is_absolute_url",
+    "normalize_url",
+    "parse_url_safe",
+    "sanitize_filename_part",
+    "validate_url",
+]
